% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obr_annual.R
\name{obr_annual}
\alias{obr_annual}
\title{Fetch Outturn and OBR forecast of inflation measures}
\usage{
obr_annual(indicator, long = FALSE)
}
\arguments{
\item{indicator}{OBR indicator, e.g., "CPI"}

\item{long}{Long (T) or wide (F) format}
}
\value{
A \code{data.frame} containing a time series
}
\description{
function
}
\examples{
obr_annual("CPI", long = FALSE)
obr_annual(c("CPI", "RPI"), long = TRUE)

}
