% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obr_download}
\alias{obr_download}
\title{Download the latest OBR data}
\usage{
obr_download()
}
\value{
The location of the downloaded Excel workbook
}
\description{
Download the latest Excel workbook of OBR data
}
