% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obr_meta.R
\name{obr_meta}
\alias{obr_meta}
\title{Fetch available indicators}
\usage{
obr_meta()
}
\value{
A \code{data.frame} containing available indicators
}
\description{
Function to check available indicators
}
\examples{
obr_annual("CPI")

}
